requirements.txt
fastapi==0.78.0
uvicorn==0.18.2
sqlalchemy==1.4.35
alembic==1.7.7
python-dotenv==0.20.0
Jinja2==3.1.2
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
pydantic[email]
python-multipart